backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.2+cpu

fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
pydantic==2.12.3
pydantic-settings==2.10.1
numpy==1.26.4
pandas==2.1.4
scikit-learn==1.7.1
joblib==1.5.2
requests==2.31.0
python-multipart==0.0.6
httpx
PyMuPDF==1.24.0
pytesseract==0.3.13
Pillow==10.0.0
python-docx==1.1.0
transformers==4.41.2
sentence-transformers==2.7.0
faiss-cpu==1.7.4
asyncpg==0.29.0